ZMQ adapter: TCP connect support

+++ package/zmq_adapter/src/zmq_adapter.c
@@ -37,7 +37,8 @@ typedef enum {
   IO_INVALID,
   IO_STDIO,
   IO_FILE,
-  IO_TCP_LISTEN
+  IO_TCP_LISTEN,
+  IO_TCP_CONNECT
 } io_mode_t;
 
 typedef enum {
@@ -76,18 +77,18 @@ static const char *zmq_req_addr = NULL;
 static const char *zmq_rep_addr = NULL;
 static const char *file_path = NULL;
 static int tcp_listen_port = -1;
+static const char *tcp_connect_addr = NULL;
 
-static void debug_printf(const char *msg, ...)
-{
-  if (!debug) {
-    return;
-  }
-
-  va_list ap;
-  va_start(ap, msg);
-  vfprintf(stderr, msg, ap);
-  va_end(ap);
-}
+static pid_t pub_pid = -1;
+static pid_t sub_pid = -1;
+static pid_t req_pid = -1;
+static pid_t rep_pid = -1;
+static pid_t *pids[] = {
+  &pub_pid,
+  &sub_pid,
+  &req_pid,
+  &rep_pid
+};
 
 static void usage(char *command)
 {
@@ -117,6 +118,7 @@ static void usage(char *command)
   fprintf(stderr, "\t--stdio\n");
   fprintf(stderr, "\t--file <file>\n");
   fprintf(stderr, "\t--tcp-l <port>\n");
+  fprintf(stderr, "\t--tcp-c <addr>\n");
 
   fprintf(stderr, "\nMisc options\n");
   fprintf(stderr, "\t--rep-timeout <ms>\n");
@@ -132,6 +134,7 @@ static int parse_options(int argc, char *argv[])
     OPT_ID_STDIO = 1,
     OPT_ID_FILE,
     OPT_ID_TCP_LISTEN,
+    OPT_ID_TCP_CONNECT,
     OPT_ID_REP_TIMEOUT,
     OPT_ID_STARTUP_DELAY,
     OPT_ID_DEBUG,
@@ -150,6 +153,7 @@ static int parse_options(int argc, char *argv[])
     {"stdio",             no_argument,       0, OPT_ID_STDIO},
     {"file",              required_argument, 0, OPT_ID_FILE},
     {"tcp-l",             required_argument, 0, OPT_ID_TCP_LISTEN},
+    {"tcp-c",             required_argument, 0, OPT_ID_TCP_CONNECT},
     {"rep-timeout",       required_argument, 0, OPT_ID_REP_TIMEOUT},
     {"startup-delay",     required_argument, 0, OPT_ID_STARTUP_DELAY},
     {"filter-in",         required_argument, 0, OPT_ID_FILTER_IN},
@@ -179,6 +183,12 @@ static int parse_options(int argc, char *argv[])
       case OPT_ID_TCP_LISTEN: {
         io_mode = IO_TCP_LISTEN;
         tcp_listen_port = strtol(optarg, NULL, 10);
+      }
+      break;
+
+      case OPT_ID_TCP_CONNECT: {
+        io_mode = IO_TCP_CONNECT;
+        tcp_connect_addr = optarg;
       }
       break;
 
@@ -294,19 +304,12 @@ static int parse_options(int argc, char *argv[])
   return 0;
 }
 
-static void sigchld_handler(int signum)
-{
-  int saved_errno = errno;
-  while (waitpid(-1, NULL, WNOHANG) > 0) {
-    ;
-  }
-  errno = saved_errno;
-}
-
 static void terminate_handler(int signum)
 {
-  /* Send this signal to the entire process group */
-  killpg(0, signum);
+  /* If this is the parent, send this signal to the entire process group */
+  if (getpid() == getpgid(0)) {
+    killpg(0, signum);
+  }
 
   /* Exit */
   _exit(EXIT_SUCCESS);
@@ -790,13 +793,48 @@ static void io_loop_reqrep(handle_t *req_handle, handle_t *rep_handle)
   debug_printf("io loop end\n");
 }
 
+static int pid_wait_check(pid_t *pid, pid_t wait_pid)
+{
+  if (*pid > 0) {
+    if (*pid == wait_pid) {
+      *pid = -1;
+      return 0;
+    }
+  }
+
+  return -1;
+}
+
+static void pid_terminate(pid_t *pid)
+{
+  if (*pid > 0) {
+    if (kill(*pid, SIGTERM) != 0) {
+      syslog(LOG_ERR, "error terminating pid %d", *pid);
+    }
+    *pid = -1;
+  }
+}
+
+void debug_printf(const char *msg, ...)
+{
+  if (!debug) {
+    return;
+  }
+
+  va_list ap;
+  va_start(ap, msg);
+  vfprintf(stderr, msg, ap);
+  va_end(ap);
+}
+
 void io_loop_start(int read_fd, int write_fd)
 {
   switch (zsock_mode) {
     case ZSOCK_PUBSUB: {
 
       if (zmq_pub_addr != NULL) {
-        if (fork() == 0) {
+        pid_t pid = fork();
+        if (pid == 0) {
           /* child process */
           zsock_t *pub = zsock_start(ZMQ_PUB);
           if (pub == NULL) {
@@ -823,11 +861,15 @@ void io_loop_start(int read_fd, int write_fd)
           handle_deinit(&pub_handle);
           handle_deinit(&fd_handle);
           exit(EXIT_SUCCESS);
+        } else {
+          /* parent process */
+          pub_pid = pid;
         }
       }
 
       if (zmq_sub_addr != NULL) {
-        if (fork() == 0) {
+        pid_t pid = fork();
+        if (pid == 0) {
           /* child process */
           zsock_t *sub = zsock_start(ZMQ_SUB);
           if (sub == NULL) {
@@ -854,6 +896,9 @@ void io_loop_start(int read_fd, int write_fd)
           handle_deinit(&sub_handle);
           handle_deinit(&fd_handle);
           exit(EXIT_SUCCESS);
+        } else {
+          /* parent process */
+          sub_pid = pid;
         }
       }
 
@@ -862,7 +907,8 @@ void io_loop_start(int read_fd, int write_fd)
 
     case ZSOCK_REQ: {
 
-      if (fork() == 0) {
+      pid_t pid = fork();
+      if (pid == 0) {
         /* child process */
         zsock_t *req = zsock_start(ZMQ_REQ);
         if (req == NULL) {
@@ -888,6 +934,9 @@ void io_loop_start(int read_fd, int write_fd)
         handle_deinit(&req_handle);
         handle_deinit(&fd_handle);
         exit(EXIT_SUCCESS);
+      } else {
+        /* parent process */
+        req_pid = pid;
       }
 
     }
@@ -895,7 +944,8 @@ void io_loop_start(int read_fd, int write_fd)
 
     case ZSOCK_REP: {
 
-      if (fork() == 0) {
+      pid_t pid = fork();
+      if (pid == 0) {
         /* child process */
         zsock_t *rep = zsock_start(ZMQ_REP);
         if (rep == NULL) {
@@ -921,6 +971,9 @@ void io_loop_start(int read_fd, int write_fd)
         handle_deinit(&rep_handle);
         handle_deinit(&fd_handle);
         exit(EXIT_SUCCESS);
+      } else {
+        /* parent process */
+        rep_pid = pid;
       }
 
     }
@@ -928,6 +981,55 @@ void io_loop_start(int read_fd, int write_fd)
 
     default:
       break;
+  }
+}
+
+void io_loop_wait(void)
+{
+  while (1) {
+    int ret = waitpid(-1, NULL, 0);
+    if ((ret == -1) && (errno == EINTR)) {
+      /* Retry if interrupted */
+      continue;
+    } else if (ret >= 0) {
+      /* Continue on success */
+      continue;
+    } else {
+      /* Break on error */
+      break;
+    }
+  }
+}
+
+void io_loop_wait_one(void)
+{
+  while (1) {
+    pid_t pid = waitpid(-1, NULL, 0);
+    if ((pid == -1) && (errno == EINTR)) {
+      /* Retry if interrupted */
+      continue;
+    } else if (pid >= 0) {
+      int i;
+      for (i = 0; i < sizeof(pids) / sizeof(pids[0]); i++) {
+        if (pid_wait_check(&pub_pid, pid) == 0) {
+          /* Return if a child from the list was terminated */
+          return;
+        }
+      }
+      /* Retry if the child was not in the list */
+      continue;
+    } else {
+      /* Return on error */
+      return;
+    }
+  }
+}
+
+void io_loop_terminate(void)
+{
+  int i;
+  for (i = 0; i < sizeof(pids) / sizeof(pids[0]); i++) {
+    pid_terminate(pids[i]);
   }
 }
 
@@ -959,16 +1061,6 @@ int main(int argc, char *argv[])
 
   signal(SIGPIPE, SIG_IGN); /* Allow write to return an error */
 
-  /* Set up SIGCHLD handler */
-  struct sigaction sigchld_sa;
-  sigchld_sa.sa_handler = sigchld_handler;
-  sigemptyset(&sigchld_sa.sa_mask);
-  sigchld_sa.sa_flags = SA_NOCLDSTOP;
-  if (sigaction(SIGCHLD, &sigchld_sa, NULL) != 0) {
-    syslog(LOG_ERR, "error setting up sigchld handler");
-    exit(EXIT_FAILURE);
-  }
-
   /* Set up handler for signals which should terminate the program */
   struct sigaction terminate_sa;
   terminate_sa.sa_handler = terminate_handler;
@@ -999,6 +1091,12 @@ int main(int argc, char *argv[])
     case IO_TCP_LISTEN: {
       extern int tcp_listen_loop(int port);
       ret = tcp_listen_loop(tcp_listen_port);
+    }
+    break;
+
+    case IO_TCP_CONNECT: {
+      extern int tcp_connect_loop(const char *addr);
+      ret = tcp_connect_loop(tcp_connect_addr);
     }
     break;
